+++ include/kernel/ramfs.h
@@ -20,8 +20,8 @@
 
 #include <kernel/fs.h>
 
-int ramfs_read(struct file *file, char *buf, size_t len, unsigned long *pos);
-int ramfs_write(struct file *file, const char *buf, size_t len, unsigned long *pos);
+ssize_t ramfs_read(struct file *file, char *buf, size_t len, unsigned long *pos);
+ssize_t ramfs_write(struct file *file, const char *buf, size_t len, unsigned long *pos);
 int ramfs_readdir(struct inode *dir, struct file *file, struct dirent *dirent,
 		int count);
 int ramfs_create(struct inode *dir, const char *name, int len, int mode,
